+++ shell/browser/ui/devtools_ui_theme_data_source.h
@@ -9,35 +9,6 @@
 #include "content/public/browser/url_data_source.h"
 
 namespace electron {
-
-// A BundledDataSource implementation that handles devtools://devtools/
-// requests.
-class BundledDataSource : public content::URLDataSource {
- public:
-  BundledDataSource() = default;
-  ~BundledDataSource() override = default;
-
-  // disable copy
-  BundledDataSource(const BundledDataSource&) = delete;
-  BundledDataSource& operator=(const BundledDataSource&) = delete;
-
-  std::string GetSource() override;
-
-  void StartDataRequest(const GURL& url,
-                        const content::WebContents::Getter& wc_getter,
-                        GotDataCallback callback) override;
-
- private:
-  std::string GetMimeType(const GURL& url) override;
-
-  bool ShouldAddContentSecurityPolicy() override;
-  bool ShouldDenyXFrameOptions() override;
-  bool ShouldServeMimeTypeAsContentTypeHeader() override;
-
-  void StartBundledDataRequest(const std::string& path,
-                               GotDataCallback callback);
-};
-
 // A ThemeDataSource implementation that handles devtools://theme/
 // requests.
 class ThemeDataSource : public content::URLDataSource {
@@ -63,5 +34,4 @@ class ThemeDataSource : public content::URLDataSource {
 };
 
 }  // namespace electron
-
-#endif  // ELECTRON_SHELL_BROWSER_UI_DEVTOOLS_UI_DATA_SOURCE_H_
+#endif
